# src/RadixJobServer/src/RadixJobServer/wwwroot/openapi-original.json
{
  "openapi" : "3.0.1",
  "info" : {
    "contact" : {
      "url" : "https://equinor.slack.com/messages/CBKM6N2JY"
    },
    "description" : "This is the API Server for the Radix job scheduler server.",
    "title" : "Radix job scheduler server.",
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "/api/v1"
  } ],
  "paths" : {
...
    "/jobs" : {
      "post" : {
        "operationId" : "createJob",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/JobScheduleDescription"
              }
            }
          },
          "description" : "Job to create",
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/JobStatus"
                }
              }
            },
            "description" : "Successful create job"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Status"
                }
              }
            },
            "description" : "Bad request"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Status"
                }
              }
            },
            "description" : "Not found"
          },
          "422" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Status"
                }
              }
            },
            "description" : "Invalid data in request"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Status"
                }
              }
            },
            "description" : "Internal server error"
          }
        },
        "summary" : "Create job",
        "tags" : [ "Job" ],
        "x-codegen-request-body-name" : "jobCreation"
      }
    },
    "/jobs/" : {
      "get" : {
        "operationId" : "getJobs",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/JobStatus"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "Successful get jobs"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Status"
                }
              }
            },
            "description" : "Internal server error"
          }
        },
        "summary" : "Gets jobs",
        "tags" : [ "Job" ]
      }
    },
    "/jobs/{jobName}" : {
      "delete" : {
        "operationId" : "deleteJob",
        "parameters" : [ {
          "description" : "Name of job",
          "in" : "path",
          "name" : "jobName",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Status"
                }
              }
            },
            "description" : "Successful delete job"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Status"
                }
              }
            },
            "description" : "Not found"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Status"
                }
              }
            },
            "description" : "Internal server error"
          }
        },
        "summary" : "Delete job",
        "tags" : [ "Job" ]
      },
      "get" : {
        "operationId" : "getJob",
        "parameters" : [ {
          "description" : "Name of job",
          "in" : "path",
          "name" : "jobName",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/JobStatus"
                }
              }
            },
            "description" : "Successful get job"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Status"
                }
              }
            },
            "description" : "Not found"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Status"
                }
              }
            },
            "description" : "Internal server error"
          }
        },
        "summary" : "Gets job",
        "tags" : [ "Job" ]
      }
    },
    "/jobs/{jobName}/stop" : {
      "post" : {
        "operationId" : "stopJob",
        "parameters" : [ {
          "description" : "Name of job",
          "in" : "path",
          "name" : "jobName",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Status"
                }
              }
            },
            "description" : "Successful delete job"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Status"
                }
              }
            },
            "description" : "Bad request"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Status"
                }
              }
            },
            "description" : "Not found"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Status"
                }
              }
            },
            "description" : "Internal server error"
          }
        },
        "summary" : "Stop job",
        "tags" : [ "Job" ]
      }
    }
  },
  "components" : {
    "schemas" : {
...
      "JobScheduleDescription" : {
        "description" : "JobScheduleDescription holds description about scheduling job",
        "example" : {
          "jobId" : "'job1'",
          "node" : "{}",
          "backoffLimit" : 1,
          "payload" : "{'data':'value'}",
          "timeLimitSeconds" : 5,
          "imageTagName" : "imageTagName",
          "resources" : "{}"
        },
        "properties" : {
          "backoffLimit" : {
            "description" : "BackoffLimit defines attempts to restart job if it fails. Corresponds to BackoffLimit in K8s.",
            "format" : "int32",
            "type" : "integer",
            "x-go-name" : "BackoffLimit"
          },
          "imageTagName" : {
            "description" : "ImageTagName defines the image tag name to use for the job image",
            "type" : "string",
            "x-go-name" : "ImageTagName"
          },
          "jobId" : {
            "description" : "JobId Optional ID of a job",
            "example" : "'job1'",
            "type" : "string",
            "x-go-name" : "JobId"
          },
          "node" : {
            "description" : "Node defines node attributes, where container should be scheduled",
            "type" : "object",
            "x-go-name" : "Node"
          },
          "payload" : {
            "description" : "Payload holding json data to be mapped to component",
            "example" : "{'data':'value'}",
            "type" : "string",
            "x-go-name" : "Payload"
          },
          "resources" : {
            "description" : "Resource describes the compute resource requirements.",
            "type" : "object",
            "x-go-name" : "Resources"
          },
          "timeLimitSeconds" : {
            "description" : "TimeLimitSeconds defines maximum job run time. Corresponds to ActiveDeadlineSeconds in K8s.",
            "format" : "int64",
            "type" : "integer",
            "x-go-name" : "TimeLimitSeconds"
          }
        },
        "type" : "object",
        "x-go-package" : "github.com/equinor/radix-job-scheduler/models/common"
      },
      "JobStatus" : {
        "description" : "JobStatus holds general information about job status",
        "example" : {
          "batchName" : "'batch1'",
          "DeploymentName" : "DeploymentName",
          "restart" : "restart",
          "created" : "2006-01-02T15:04:05Z",
          "started" : "2006-01-02T15:04:05Z",
          "failed" : 6,
          "batchId" : "'batch-id-1'",
          "message" : "\"Error occurred\"",
          "podStatuses" : [ {
            "image" : "radixdev.azurecr.io/app-server:cdgkg",
            "reason" : "reason",
            "imageId" : "radixdev.azurecr.io/app-server@sha256:d40cda01916ef63da3607c03785efabc56eb2fc2e0dab0726b1a843e9ded093f",
            "replicaStatus" : {
              "status" : "Running"
            },
            "restartCount" : 5,
            "created" : "2006-01-02T15:04:05Z",
            "containerStarted" : "2006-01-02T15:04:05Z",
            "statusMessage" : "statusMessage",
            "name" : "server-78fc8857c4-hm76l",
            "exitCode" : 1,
            "podIndex" : 5,
            "startTime" : "2006-01-02T15:04:05Z",
            "endTime" : "2006-01-02T15:04:05Z"
          }, {
            "image" : "radixdev.azurecr.io/app-server:cdgkg",
            "reason" : "reason",
            "imageId" : "radixdev.azurecr.io/app-server@sha256:d40cda01916ef63da3607c03785efabc56eb2fc2e0dab0726b1a843e9ded093f",
            "replicaStatus" : {
              "status" : "Running"
            },
            "restartCount" : 5,
            "created" : "2006-01-02T15:04:05Z",
            "containerStarted" : "2006-01-02T15:04:05Z",
            "statusMessage" : "statusMessage",
            "name" : "server-78fc8857c4-hm76l",
            "exitCode" : 1,
            "podIndex" : 5,
            "startTime" : "2006-01-02T15:04:05Z",
            "endTime" : "2006-01-02T15:04:05Z"
          } ],
          "jobId" : "'job1'",
          "ended" : "2006-01-02T15:04:05Z",
          "name" : "calculator",
          "updated" : "2006-01-02T15:04:05Z",
          "status" : "Waiting"
        },
        "properties" : {
          "DeploymentName" : {
            "description" : "DeploymentName for this batch",
            "type" : "string"
          },
          "batchId" : {
            "description" : "Defines a user defined ID of the batch.",
            "example" : "'batch-id-1'",
            "type" : "string",
            "x-go-name" : "BatchId"
          },
          "batchName" : {
            "description" : "BatchName Optional Batch ID of a job",
            "example" : "'batch1'",
            "type" : "string",
            "x-go-name" : "BatchName"
          },
          "created" : {
            "description" : "Created timestamp",
            "example" : "2006-01-02T15:04:05Z",
            "type" : "string",
            "x-go-name" : "Created"
          },
          "ended" : {
            "description" : "Ended timestamp",
            "example" : "2006-01-02T15:04:05Z",
            "type" : "string",
            "x-go-name" : "Ended"
          },
          "failed" : {
            "description" : "The number of times the container for the job has failed.\n+optional",
            "format" : "int32",
            "type" : "integer",
            "x-go-name" : "Failed"
          },
          "jobId" : {
            "description" : "JobId Optional ID of a job",
            "example" : "'job1'",
            "type" : "string",
            "x-go-name" : "JobId"
          },
          "message" : {
            "description" : "Message, if any, of the job",
            "example" : "\"Error occurred\"",
            "type" : "string",
            "x-go-name" : "Message"
          },
          "name" : {
            "description" : "Name of the job",
            "example" : "calculator",
            "type" : "string",
            "x-go-name" : "Name"
          },
          "podStatuses" : {
            "description" : "PodStatuses for each pod of the job",
            "items" : {
              "$ref" : "#/components/schemas/PodStatus"
            },
            "type" : "array",
            "x-go-name" : "PodStatuses"
          },
          "restart" : {
            "description" : "Timestamp of the job restart, if applied.\n+optional",
            "type" : "string",
            "x-go-name" : "Restart"
          },
          "started" : {
            "description" : "Started timestamp",
            "example" : "2006-01-02T15:04:05Z",
            "type" : "string",
            "x-go-name" : "Started"
          },
          "status" : {
            "description" : "Status of the job\nRunning = Job is running\nSucceeded = Job has succeeded\nFailed = Job has failed\nWaiting = Job is waiting\nStopping = Job is stopping\nStopped = Job has been stopped\nActive = Job is active\nCompleted = Job is completed",
            "enum" : [ "Running", "Succeeded", "Failed", "Waiting", "Stopping", "Stopped", "Active", "Completed" ],
            "example" : "Waiting",
            "type" : "string",
            "x-go-name" : "Status"
          },
          "updated" : {
            "description" : "Updated timestamp when the status was updated",
            "example" : "2006-01-02T15:04:05Z",
            "type" : "string",
            "x-go-name" : "Updated"
          }
        },
        "required" : [ "created", "name" ],
        "type" : "object",
        "x-go-package" : "github.com/equinor/radix-job-scheduler/models/v1"
      },
      "PodStatus" : {
        "example" : {
          "image" : "radixdev.azurecr.io/app-server:cdgkg",
          "reason" : "reason",
          "imageId" : "radixdev.azurecr.io/app-server@sha256:d40cda01916ef63da3607c03785efabc56eb2fc2e0dab0726b1a843e9ded093f",
          "replicaStatus" : {
            "status" : "Running"
          },
          "restartCount" : 5,
          "created" : "2006-01-02T15:04:05Z",
          "containerStarted" : "2006-01-02T15:04:05Z",
          "statusMessage" : "statusMessage",
          "name" : "server-78fc8857c4-hm76l",
          "exitCode" : 1,
          "podIndex" : 5,
          "startTime" : "2006-01-02T15:04:05Z",
          "endTime" : "2006-01-02T15:04:05Z"
        },
        "properties" : {
          "containerStarted" : {
            "description" : "Container started timestamp",
            "example" : "2006-01-02T15:04:05Z",
            "type" : "string",
            "x-go-name" : "ContainerStarted"
          },
          "created" : {
            "description" : "Created timestamp",
            "example" : "2006-01-02T15:04:05Z",
            "type" : "string",
            "x-go-name" : "Created"
          },
          "endTime" : {
            "description" : "The time at which the batch job's pod finishedAt.",
            "example" : "2006-01-02T15:04:05Z",
            "type" : "string",
            "x-go-name" : "EndTime"
          },
          "exitCode" : {
            "description" : "Exit status from the last termination of the container",
            "format" : "int32",
            "type" : "integer",
            "x-go-name" : "ExitCode"
          },
          "image" : {
            "description" : "The image the container is running.",
            "example" : "radixdev.azurecr.io/app-server:cdgkg",
            "type" : "string",
            "x-go-name" : "Image"
          },
          "imageId" : {
            "description" : "ImageID of the container's image.",
            "example" : "radixdev.azurecr.io/app-server@sha256:d40cda01916ef63da3607c03785efabc56eb2fc2e0dab0726b1a843e9ded093f",
            "type" : "string",
            "x-go-name" : "ImageId"
          },
          "name" : {
            "description" : "Pod name",
            "example" : "server-78fc8857c4-hm76l",
            "type" : "string",
            "x-go-name" : "Name"
          },
          "podIndex" : {
            "description" : "The index of the pod in the re-starts",
            "format" : "int64",
            "type" : "integer",
            "x-go-name" : "PodIndex"
          },
          "reason" : {
            "description" : "A brief CamelCase message indicating details about why the job is in this phase",
            "type" : "string",
            "x-go-name" : "Reason"
          },
          "replicaStatus" : {
            "$ref" : "#/components/schemas/ReplicaStatus"
          },
          "restartCount" : {
            "description" : "RestartCount count of restarts of a component container inside a pod",
            "format" : "int32",
            "type" : "integer",
            "x-go-name" : "RestartCount"
          },
          "startTime" : {
            "description" : "The time at which the batch job's pod startedAt",
            "example" : "2006-01-02T15:04:05Z",
            "type" : "string",
            "x-go-name" : "StartTime"
          },
          "statusMessage" : {
            "description" : "StatusMessage provides message describing the status of a component container inside a pod",
            "type" : "string",
            "x-go-name" : "StatusMessage"
          }
        },
        "required" : [ "name" ],
        "title" : "PodStatus contains details for the current status of the job's pods.",
        "type" : "object",
        "x-go-package" : "github.com/equinor/radix-job-scheduler/models/v1"
      },
...
      "Status" : {
        "description" : "Status is a return value for calls that don't return other objects or when a request returns an error",
        "example" : {
          "reason" : "reason",
          "code" : 404,
          "message" : "job job123 is not found",
          "status" : "Failure"
        },
        "properties" : {
          "code" : {
            "description" : "Suggested HTTP return code for this status, 0 if not set.",
            "example" : 404,
            "format" : "int64",
            "type" : "integer",
            "x-go-name" : "Code"
          },
          "message" : {
            "description" : "A human-readable description of the status of this operation.",
            "example" : "job job123 is not found",
            "type" : "string",
            "x-go-name" : "Message"
          },
          "reason" : {
            "type" : "string",
            "x-go-package" : "github.com/equinor/radix-job-scheduler/models/common"
          },
          "status" : {
            "description" : "Status of the operation.\nOne of: \"Success\" or \"Failure\".",
            "example" : "Failure",
            "type" : "string",
            "x-go-name" : "Status"
          }
        },
        "type" : "object",
        "x-go-package" : "github.com/equinor/radix-job-scheduler/models/common"
      },
...
    }
  },
...
}